move formatBuildError to react-dev-utils
