+++ packages/react-dev-utils/package.json
@@ -18,13 +18,14 @@
     "crossSpawn.js",
     "eslintFormatter.js",
     "FileSizeReporter.js",
+    "formatBuildError.js",
     "formatWebpackMessages.js",
     "getProcessForPort.js",
     "inquirer.js",
     "InterpolateHtmlPlugin.js",
     "launchEditor.js",
-    "noopServiceWorkerMiddleware.js",
     "ModuleScopePlugin.js",
+    "noopServiceWorkerMiddleware.js",
     "openBrowser.js",
     "openChrome.applescript",
     "printHostingInstructions.js",
